HTBHF-2165 refactor wiremock mappings
