+++ src/test/common/wiremock/claim/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./claim')
