+++ src/test/common/wiremock/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./claim'),
+  ...require('./postcode-lookup'),
+  ...require('./wiremock')
+}
